Extract the supermod template list into its own component
The grouped template list lived inline in SunshineUserMessages, which meant it
could only ever serve message templates, and made that file the collision point
for anything touching either concern.

GroupedModerationTemplateList takes a collectionName and owns its own query,
grouping, search, collapse state and drag-to-regroup — no behaviour change for
the DM tab. The "new template" button moves in with it, so the component is
self-contained rather than needing its caller to render half of it;
ModerationTemplatesForm gains initialCollectionName so a new template defaults
to the list it was created from.
